KYLIN-5403 fix potential deadlock of class loading of RoutingRule.java
