+++ src/query-common/src/main/java/org/apache/kylin/query/routing/QueryRouter.java
@@ -26,6 +26,11 @@ import org.apache.kylin.metadata.model.FunctionDesc;
 import org.apache.kylin.metadata.realization.IRealization;
 import org.apache.kylin.query.relnode.OLAPContext;
 import org.apache.kylin.query.relnode.OLAPContextProp;
+import org.apache.kylin.query.routing.rules.PartitionPruningRule;
+import org.apache.kylin.query.routing.rules.RealizationSortRule;
+import org.apache.kylin.query.routing.rules.RemoveBlackoutRealizationsRule;
+import org.apache.kylin.query.routing.rules.RemoveUncapableRealizationsRule;
+import org.apache.kylin.query.routing.rules.SegmentPruningRule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -42,6 +47,45 @@ public class QueryRouter {
     private QueryRouter() {
     }
 
+    private static final List<RoutingRule> LAYOUT_CHOOSING_RULES = Lists.newLinkedList();
+
+    static {
+        LAYOUT_CHOOSING_RULES.add(new RemoveBlackoutRealizationsRule());
+        LAYOUT_CHOOSING_RULES.add(new SegmentPruningRule());
+        LAYOUT_CHOOSING_RULES.add(new PartitionPruningRule());
+        LAYOUT_CHOOSING_RULES.add(new RemoveUncapableRealizationsRule());
+        LAYOUT_CHOOSING_RULES.add(new RealizationSortRule());
+    }
+
+    public static void applyRules(List<Candidate> candidates) {
+        for (RoutingRule rule : LAYOUT_CHOOSING_RULES) {
+            String before = getPrintableText(candidates);
+            rule.apply(candidates);
+            String after = getPrintableText(candidates);
+            if (!before.equals(after)) {
+                logger.info("Applying rule: {}, realizations before:{}, realizations after: {}", rule, before, after);
+            }
+        }
+    }
+
+    public static String getPrintableText(List<Candidate> candidates) {
+        StringBuilder sb = new StringBuilder();
+        sb.append("[");
+        for (Candidate candidate : candidates) {
+            IRealization r = candidate.realization;
+            sb.append(r.getCanonicalName());
+            sb.append(",");
+        }
+        if (sb.charAt(sb.length() - 1) != '[')
+            sb.deleteCharAt(sb.length() - 1);
+        sb.append("]");
+        return sb.toString();
+    }
+
+    public static void removeRule(RoutingRule rule) {
+        LAYOUT_CHOOSING_RULES.removeIf(r -> r.getClass() == rule.getClass());
+    }
+
     public static Candidate selectRealization(OLAPContext olapContext, IRealization realization,
             Map<String, String> aliasMap) {
         if (!realization.isReady()) {
@@ -56,7 +100,7 @@ public class QueryRouter {
         List<Candidate> originCandidates = Lists.newArrayList(candidates);
 
         // rule based realization selection, rules might reorder realizations or remove specific realization
-        RoutingRule.applyRules(candidates);
+        QueryRouter.applyRules(candidates);
 
         collectIncapableReason(olapContext, originCandidates);
         if (candidates.isEmpty()) {
@@ -66,7 +110,7 @@ public class QueryRouter {
         Candidate chosen = candidates.get(0);
         chosen.setRewrittenCtx(preserveRewriteProps(olapContext));
         logger.info("The realizations remaining: {}, and the final chosen one for current olap context {} is {}",
-                RoutingRule.getPrintableText(candidates), olapContext.id, chosen.realization.getCanonicalName());
+                QueryRouter.getPrintableText(candidates), olapContext.id, chosen.realization.getCanonicalName());
         return chosen;
     }
